make package for retrieverapp
