+++ .github/workflows/test.yml
@@ -41,7 +41,7 @@ jobs:
           NCBI_API_EMAIL: ${{ secrets.NCBI_API_EMAIL }}
           NCBI_API_KEY: ${{ secrets.NCBI_API_KEY }}
         run: |
-          python get_data.py -grants example_grants.txt
+          retriever_get -grants example_grants.txt
           mkdir output
           cp -r retriever_app.html JSON_data custom_text.json output/
       - name: Archive output
