+++ pyproject.toml
@@ -1,0 +1,43 @@
+[build-system]
+requires = ["setuptools>=42", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+include-package-data = true
+
+[tool.setuptools.packages.find]
+namespaces = true
+where = ["src"]
+
+
+[project]
+name = "RetrieverApp"
+version = "0.1.0"
+description = "Retrieve open accessible resources from list of publications or grant numbers"
+authors = [{name = "Philippa Doherty", email = "Philippa.Doherty@RoswellPark.org"}]
+license = {file = "LICENSE.txt"}
+readme = "README.md"
+requires-python = ">=3.6"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
+dependencies = [
+    "requests~=2.26",
+    "pandas~=1.3",
+    "biopython~=1.78",
+    "beautifulsoup4~=4.10",
+    "xmltodict",
+    "openpyxl",
+]
+
+[project.urls]
+Homepage = "http://somewebsite.org"
+Repository = "http://somewebsite.org"
+
+# define the console scripts
+[project.entry-points.console_scripts]
+retriever_get = "retrieverapp.get_data:main"
+retriever_refresh = "retrieverapp.refresh_data:main"
